scanner file
